requirements.txt
flask
mysql-connector-python
gunicorn
bcrypt